requirements.txt
requests==2.31.0
pandas==2.0.3
pytest==7.4.0
python-dotenv==1.0.0